refactor: split workflow expression boot checks
